% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dlunif.R, R/plunif.R, R/qlunif.R, R/rlunif.R
\name{dlunif}
\alias{dlunif}
\alias{plunif}
\alias{qlunif}
\alias{rlunif}
\title{The Log Uniform Distribution}
\usage{
dlunif(x, min, max, base = exp(1))

plunif(q, min, max, base = exp(1))

qlunif(p, min, max, base = exp(1))

rlunif(n, min, max, base = exp(1))
}
\arguments{
\item{x}{Vector of quantiles.}

\item{min}{Lower limit of the distribution, in raw (untransformed in
log-space) values. Negative values will give warning.}

\item{max}{Upper limit of the distribution, in raw (untransformed in
log-space) values. Negative values will give warning.}

\item{base}{The base to which logarithms are computed. Defaults to
  \code{e=exp(1)}. Must be a positive or complex number.}

\item{q}{Vector of quantiles.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
\code{dlunif} gives the density, \code{plunif} gives the distribution
  function, \code{qlunif} gives the quantile function, and \code{rlunif}
  generates random numbers.
}
\description{
Density, distribution function, quantile function and random generation for
the log uniform distribution in the interval from \code{min} to \code{max}.
Parameters must be raw values (untransformed in log-space) so that can be
log-transformed using specified \code{base}.
}
\details{
The log uniform (or loguniform or log-uniform) distribution behaves
  as a uniform distribution when transformed to log-space.
}
\note{
Parameters \code{min, max} must be provided as raw (untransformed in
  log-space) values so that can be log-transformed using \code{base}.
}
\examples{
plot(1:100, dlunif(1:100, 1, 100), type="l", main="Loguniform density")
plot(log(1:100), dlunif(log(1:100), log(1), log(100)), type="l",
     main="Loguniform density")

plot(1:100, plunif(1:100, 1, 100), type="l", main="Loguniform cumulative")

plot(qlunif(ppoints(100), 1, 100), type="l", main="Loguniform quantile")

hist(rlunif(1000, 1, 1000), main="random loguniform sample")
hist(log(rlunif(10000, 1, 1000)), main="random loguniform sample")
hist(log(rlunif(10000, 1, 1000, base=10), base=10), main="random loguniform sample")
}
\author{
Steve Wang \email{scwang@swarthmore.edu}
}
\seealso{
\code{\link[stats]{Distributions}} for other standard distributions
}

