% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test_stat.R
\name{ks_test_stat}
\alias{ks_test_stat}
\title{Internal KScorrect Function.}
\usage{
ks_test_stat(x, y, ...)
}
\arguments{
\item{x}{a numeric vector of data values.}

\item{y}{a character string naming a cumulative distribution function or an
actual cumulative distribution function such as pnorm. Only continuous CDFs
are valid. See /code{LcKS} for accepted functions.}

\item{...}{parameters of the distribution specified (as a character string)
by y.}
}
\description{
Internal function not intended to be called directly by users.
}
\details{
Simplified and faster \code{\link{ks.test}} function that calculates just the
two-sided test statistic D.
}
\note{
Calculating the Kolmogorov-Smirnov test statistic D by itself is faster
  than calculating the other ouput that that function produces.
}
\seealso{
\code{\link[stats]{ks.test}}
}
