% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KScorrect.R
\docType{package}
\name{KScorrect-package}
\alias{KScorrect}
\alias{KScorrect-package}
\title{KScorrect: Lilliefors-Corrected Kolmogorov-Smirnoff Goodness-Of-Fit Tests}
\description{
Implements the Lilliefors-corrected Kolmogorov-Smirnoff test for use in
goodness-of-fit tests.
}
\details{
KScorrect implements the Lilliefors-corrected Kolmogorov-Smirnoff test for
use in goodness-of-fit tests, suitable when population parameters are unknown
and must be estimated by sample statistics. It uses a resampling algorithm to
estimate p-values from the resampling distribution. Coded to complement
\code{\link[stats]{ks.test}}, it can be used with a variety of continuous
distributions, including normal, lognormal, univariate mixtures of normals,
uniform, loguniform, exponential, gamma, and Weibull distributions.

Functions to generate random numbers and calculate density, distribution, and
quantile functions are provided for use with the loguniform and mixture
distributions.
}
\examples{
# Get the package version and citation of KScorrect
packageVersion("KScorrect")
citation("KScorrect")

x <- runif(200)
l <- LcKS(x, cdf="pnorm", nreps=999)
hist(l$D.resample)
abline(v = l$D.obs, lty = 2)
print(l, max=50)  # Just print first 50 resampled statistics
# Approximate p-value (usually) << 0.05

# Confirmation uncorrected version has biased type-I error as a one-sample
#   test using sample statistics for parameters:
ks.test(x, "pnorm", mean(x), sd(x))   # p-value always larger, (usually) > 0.05

x <- rlunif(200, min=1, max=500) # random loguniform sample
l <- LcKS(x, cdf="plnorm")
l$p.value      # Approximate p-value: (usually) << 0.05
}
\author{
Phil Novack-Gottshall \email{pnovack-gottshall@ben.edu}

Steve C. Wang \email{scwang@swarthmore.edu}
}

