% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlunif.R, R/plunif.R, R/qlunif.R, R/rlunif.R
\name{dlunif}
\alias{dlunif}
\alias{plunif}
\alias{qlunif}
\alias{rlunif}
\title{The Log Uniform Distribution}
\usage{
dlunif(x, min, max, base = exp(1))

plunif(q, min, max, base = exp(1))

qlunif(p, min, max, base = exp(1))

rlunif(n, min, max, base = exp(1))
}
\arguments{
\item{x}{Vector of quantiles.}

\item{min}{Lower limit of the distribution, in raw (not log-transformed)
values. Negative values will give warning.}

\item{max}{Upper limit of the distribution, in raw (not log-transformed)
values. Negative values will give warning.}

\item{base}{The base to which logarithms are computed. Defaults to
\code{e=exp(1)}. Must be a positive number.}

\item{q}{Vector of quantiles.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
\code{dlunif} gives the density, \code{plunif} gives the distribution
  function, \code{qlunif} gives the quantile function, and \code{rlunif}
  generates random numbers.
}
\description{
Density, distribution function, quantile function and random generation for
the log uniform distribution in the interval from \code{min} to \code{max}.
Parameters must be raw values (not log-transformed) and will be
log-transformed using specified \code{base}.
}
\details{
A log uniform (or loguniform or log-uniform) random variable has a
  uniform distribution when log-transformed.
}
\note{
Parameters \code{min, max} must be provided as raw (not
  log-transformed) values and will be log-transformed using \code{base}. In
  other words, when log-transformed, a log uniform random variable with
  parameters \code{min=a} and \code{max=b} is uniform over the interval from
  \code{log(a)} to \code{log(b)}.
}
\examples{
plot(1:100, dlunif(1:100, exp(1), exp(10)), type="l", main="Loguniform density")
plot(log(1:100), dlunif(log(1:100), log(1), log(10)), type="l",
     main="Loguniform density")

plot(1:100, plunif(1:100, exp(1), exp(10)), type="l", main="Loguniform cumulative")
plot(qlunif(ppoints(100), exp(1), exp(10)), type="l", main="Loguniform quantile")

hist(rlunif(1000, exp(1), exp(10)), main="random loguniform sample")
hist(log(rlunif(10000, exp(1), exp(10))), main="random loguniform sample")
hist(log(rlunif(10000, exp(1), exp(10), base=10), base=10), main="random loguniform sample")

}
\seealso{
\code{\link[stats]{Distributions}} for other standard distributions
}
\author{
Steve Wang \email{scwang@swarthmore.edu}
}
